requirement.txt
dill==0.2.8.2
kafka-python==1.4.4
pymongo==3.7.2
